+++ main/flashnvm.h
@@ -14,7 +14,7 @@
 void init_nvs() ;
 void save_base_distance(float );
 void save_base_distance(float);
-float load_base_distance();
+int load_base_distance();
 void save_calibration_status(bool);
 bool load_calibration_status();
 void check_first_time_calibration();
